Design: transfer ImageTab function button #39
